Update donations with the user API, refactor some casing in the message sender
